Add files via upload
The Mushroom Classification Project is a machine learning application designed to classify mushrooms into edible and poisonous categories based on various features. This project leverages a dataset containing various attributes of mushrooms to build a predictive model, aiming to assist in safe mushroom foraging and consumption. The primary goal is to provide a tool that can help differentiate between safe and dangerous mushroom species, reducing the risk of accidental poisoning.