Add files via upload
Heart disease is one of the leading causes of mortality worldwide. Early detection and diagnosis are crucial for effective treatment and improving patient outcomes. This project aims to develop a machine learning model to predict the likelihood of heart disease based on various clinical parameters. The goal is to leverage advanced data analysis techniques to assist healthcare professionals in making informed decisions.